Refactor BluetoothLeManager's scanning
It doesn't make sense to have a global scan job - each scan job is for a different device. If anything, we could queue up scan jobs, but that would be more work than it's worth.

This change also allows the scanner to immediately return when it finds the device, using coroutine magic (which is possibly but hopefully not broken)

We also set scan mode to low latency in order to improve performance, especially when the screen is going to be off. I have noticed that the device is often immediately detected with this mode.